refactor: 스키마 파일 분리 및 경로 변경
