+++ src/v1/apis/schemas/getUser.schema.ts
@@ -1,0 +1,8 @@
+import { createResponseSchema } from '../../common/schema/core.schema.js';
+import { exceptedSensitiveFields, userSchema } from './users.schema.js';
+
+export const getUserParamsSchema = userSchema.pick({
+  id: true,
+});
+
+export const getUserResponseSchema = createResponseSchema(exceptedSensitiveFields);
