+++ src/v1/apis/schemas/searchUser.schema.ts
@@ -1,0 +1,9 @@
+import { exceptedSensitiveFields, userSchema } from './users.schema.js';
+import { createResponseSchema } from '../../common/schema/core.schema.js';
+import { z } from 'zod';
+
+export const searchUserParamsSchema = userSchema.pick({
+  nickname: true,
+});
+
+export const searchUserResponseSchema = createResponseSchema(z.array(exceptedSensitiveFields));
